Rename lambda folder to thin_egress_app
