+++ Makefile
@@ -1,8 +1,8 @@
 SOURCES := \
-	lambda/var/otel-config/collector.yaml \
-	lambda/app.py \
-	lambda/tea_bumper.py \
-	lambda/update_lambda.py
+	thin_egress_app/var/otel-config/collector.yaml \
+	thin_egress_app/app.py \
+	thin_egress_app/tea_bumper.py \
+	thin_egress_app/update_lambda.py
 
 HTML_TEMPLATES := $(wildcard templates/*.html)
 MD_TEMPLATES := $(wildcard templates/*.md)
@@ -15,7 +15,7 @@ REQUIREMENTS_TXT := $(REQUIREMENTS_IN:.in=.txt)
 DIR := dist
 EMPTY := $(DIR)/empty
 # Temporary artifacts
-DIST_SOURCES := $(SOURCES:lambda/%=$(DIR)/code/%)
+DIST_SOURCES := $(SOURCES:thin_egress_app/%=$(DIR)/code/%)
 DIST_MD_RESOURCES := $(MD_TEMPLATES:%.md=$(DIR)/code/%.html)
 DIST_HTML_RESOURCES := $(HTML_TEMPLATES:%=$(DIR)/code/%)
 DIST_RESOURCES := $(DIST_HTML_RESOURCES) $(DIST_MD_RESOURCES)
@@ -119,7 +119,7 @@ $(DIST_HTML_RESOURCES): $(DIR)/code/%: %
 	cp $< $@
 
 .SECONDARY: $(DIST_SOURCES)
-$(DIST_SOURCES): $(DIR)/code/%: lambda/%
+$(DIST_SOURCES): $(DIR)/code/%: thin_egress_app/%
 	@mkdir -p $(@D)
 	cp $< $@
 	$(PYTHON) scripts/sed.py -i $@ "<BUILD_ID>" "${BUILD_ID}"
@@ -286,4 +286,4 @@ lock: $(REQUIREMENTS_TXT)
 
 .PHONY: test
 test:
-	pytest --cov=lambda --cov-report=term-missing --cov-branch tests
+	pytest --cov=thin_egress_app --cov-report=term-missing --cov-branch tests
