+++ thin_egress_app/var/otel-config/collector.yaml
